Fix calculation method and create new service for current charges model
